[Refactoring]: Seperate title and menu to another component  (#3531)
* refactor: separated the title renderer

* refactor: separated styles

* refactor: seperate types
