+++ frontend/src/container/LogDetailedView/BodyTitleRenderer.styles.ts
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+
+export const TitleWrapper = styled.span`
+	.hover-reveal {
+		visibility: hidden;
+	}
+
+	&:hover .hover-reveal {
+		visibility: visible;
+	}
+`;
