+++ frontend/src/container/LogDetailedView/LogDetailedView.types.ts
@@ -1,0 +1,5 @@
+export interface BodyTitleRendererProps {
+	title: string;
+}
+
+export type AnyObject = { [key: string]: any };
