+++ frontend/src/container/LogDetailedView/utils.tsx
@@ -1,7 +1,7 @@
-import { SettingOutlined } from '@ant-design/icons';
-import { Dropdown, Menu } from 'antd';
 import { DataNode } from 'antd/es/tree';
-import styled from 'styled-components';
+
+import BodyTitleRenderer from './BodyTitleRenderer';
+import { AnyObject } from './LogDetailedView.types';
 
 export const recursiveParseJSON = (obj: string): Record<string, unknown> => {
 	try {
@@ -15,38 +15,12 @@ export const recursiveParseJSON = (obj: string): Record<string, unknown> => {
 	}
 };
 
-const TitleWrapper = styled.span`
-	.hover-reveal {
-		visibility: hidden;
-	}
-
-	&:hover .hover-reveal {
-		visibility: visible;
-	}
-`;
-
 export function jsonToDataNodes(
 	json: Record<string, unknown>,
 	parentKey = '',
 ): DataNode[] {
 	return Object.entries(json).map(([key, value]) => {
 		const nodeKey = parentKey ? `${parentKey}.${key}` : key;
-
-		const menu = (
-			<Menu>
-				<Menu.Item key="0">Option 1</Menu.Item>
-				<Menu.Item key="1">Option 2</Menu.Item>
-			</Menu>
-		);
-
-		const title = (
-			<TitleWrapper>
-				{key}
-				<Dropdown overlay={menu} trigger={['click']}>
-					<SettingOutlined style={{ marginLeft: 8 }} className="hover-reveal" />
-				</Dropdown>
-			</TitleWrapper>
-		);
 
 		if (typeof value === 'object' && value !== null) {
 			return {
@@ -57,12 +31,10 @@ export function jsonToDataNodes(
 		}
 		return {
 			key: nodeKey,
-			title,
+			title: <BodyTitleRenderer title={key} />,
 		};
 	});
 }
-
-type AnyObject = { [key: string]: any };
 
 export function flattenObject(obj: AnyObject, prefix = ''): AnyObject {
 	return Object.keys(obj).reduce((acc: AnyObject, k: string): AnyObject => {
